test: extract fakeNetworkEchoStuff
